✨  user State에 따른 room modal창 렌더링
