+++ client/src/components/right-sidebar.tsx
@@ -1,14 +1,13 @@
 /* eslint-disable object-shorthand */
-import React, { useState, useRef } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import { useRecoilState } from 'recoil';
 
+import RoomModal from '@components/room-modal';
 import roomTypeState from '@atoms/room-type';
-import DefaultButton from './styled-components/default-button';
-import RoomTypeCheckBox from './room-type-check-box';
-import AnonymousCheckBox from './anonymous-checkbox';
+import userTypeState from '@atoms/user';
 
-const RoomModal = styled.div`
+const RoomModalLayout = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: space-around;
@@ -23,73 +22,28 @@ const RoomModal = styled.div`
   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
 `;
 
-const CustomTitleForm = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-`;
-
-const TitleInputbar = styled.input`
-  border: 1px solid #58964F;
-  border-radius: 8px;
-  padding: 0;
-
-  &:focus {
-    outline: none;
-  }
-`;
-
-const TitleInputbarLabel = styled.label`
-  font-size: 12px;
-  color: #B6B6B6;
-`;
-
 function RightSideBar() {
   const [roomType] = useRecoilState(roomTypeState);
-  const [isDisabled, setIsDisabled] = useState(true);
-  const [isAnonymous, setIsAnonymous] = useState(false);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const submitButtonHandler = () => {
-    // userId 현재 아이디 가져와야함
-    const roomInfo = {
-      type: roomType,
-      title: inputRef.current?.value,
-      userId: 'dlatqdlatq',
-      isAnonymous: isAnonymous,
-    };
-    fetch(`${process.env.REACT_APP_API_URL}/api/room`, {
-      method: 'post',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(roomInfo),
+  const [user] = useRecoilState(userTypeState);
 
-    }).then((res) => console.log(res))
-      .catch((err) => console.error(err));
-  };
-
-  const inputOnChange = () => {
-    setIsDisabled(!inputRef.current?.value);
-  };
-
-  const checkboxOnChange = () => {
-    setIsAnonymous(!isAnonymous);
-  };
-
+  if (user.roomId === '') { // 방 생성 모달
+    return (
+      <RoomModalLayout>
+        <RoomModal />
+      </RoomModalLayout>
+    );
+  } if (roomType === 'closed') { // closed인 경우 특정 인원을 지정하는 화면을 만들어야함
+    return (
+      <RoomModalLayout>
+        <h1> closed!!! </h1>
+      </RoomModalLayout>
+    );
+  }
+  // 방에 들어가 있는 경우
   return (
-    <RoomModal>
-      <RoomTypeCheckBox checkBoxName="public" />
-      <RoomTypeCheckBox checkBoxName="social" />
-      <RoomTypeCheckBox checkBoxName="closed" />
-      <AnonymousCheckBox checked={isAnonymous} onChange={checkboxOnChange} />
-      <CustomTitleForm>
-        <TitleInputbarLabel>Add a Room Title</TitleInputbarLabel>
-        <TitleInputbar ref={inputRef} onChange={inputOnChange} />
-      </CustomTitleForm>
-      <DefaultButton buttonType="secondary" size="medium" onClick={submitButtonHandler} isDisabled={isDisabled}>
-        Lets Go
-      </DefaultButton>
-    </RoomModal>
+    <RoomModalLayout>
+      <h1> 방에 들어감. </h1>
+    </RoomModalLayout>
   );
 }
